backend/requirements.txt
Flask==2.2.3

python-dotenv
prettyprint

pandas
pyodbc

pytest